Fix Next.js errors
