+++ examples/nextjs-comments-notifications/src/example.ts
@@ -1,6 +1,3 @@
-import { useSearchParams } from "next/navigation";
-import { useMemo } from "react";
-
 const PLACEHOLDER_BASE_URL = "https://localhost:9999";
 const ABSOLUTE_URL_REGEX = /^[a-zA-Z][a-zA-Z\d+\-.]*?:/;
 
@@ -29,28 +26,6 @@ export function createExampleUserId(
   }
 
   return userId;
-}
-
-export function useExampleUserId() {
-  const params = useSearchParams();
-  const exampleId = params?.get("exampleId");
-  const examplePreview = params?.get("examplePreview");
-  const userId = useMemo(() => {
-    return createExampleUserId(Number(examplePreview), exampleId);
-  }, [exampleId, examplePreview]);
-
-  return userId;
-}
-
-export function useExampleRoomId(room: string) {
-  const params = useSearchParams();
-  const exampleId = params?.get("exampleId");
-  const roomId = useMemo(() => createExampleRoomId(room), [room]);
-  const exampleRoomId = useMemo(() => {
-    return exampleId ? `${roomId}-${exampleId}` : roomId;
-  }, [roomId, exampleId]);
-
-  return exampleRoomId;
 }
 
 export function setExampleId(url: string) {
